requirements/requirements.txt
requests==2.22.0
django==3.1.7
pycryptodome==3.10.1